+++ binding.gyp
@@ -136,7 +136,8 @@
         'src/create-offer-observer.cc',
         'src/datachannel.cc',
         'src/peerconnection.cc',
-        'src/events.cc',
+        'src/events/datachannel.cc',
+        'src/events/peerconnection.cc',
         'src/rtcstatsreport.cc',
         'src/rtcstatsresponse.cc',
         'src/set-local-description-observer.cc',
